Extract complex enrichDatasources into its own file
